Split out ArgsToFrontendInputsConverter.
